Working subtitles on GBA

+++ gba/data/font_sans.h
@@ -1,4 +1,4 @@
-// Converted with img2h /home/raer/SynologyDrive/Programmieren/GameBoy/GBA-image-tools/gba/data/font_sans.png --blackwhite=0.5 --outformat=bgr555 font_sans
+// Converted with img2h --blackwhite=0.5 --outformat=bgr555 data/font_sans.png data/font_sans
 // Note that the _Alignas specifier will need C11, as a workaround use __attribute__((aligned(4)))
 
 // Data is bitmap tiles, pixel format: Paletted 8-bit, color map format: XBGR1555
@@ -13,5 +13,5 @@
 extern const uint32_t FONT_SANS_DATA[FONT_SANS_DATA_SIZE];
 #define FONT_SANS_PALETTE_LENGTH 2 // # of palette entries per palette
 #define FONT_SANS_PALETTE_SIZE 4 // size of palette data
-extern const uint16_t FONT_SANS_PALETTE[FONT_SANS_PALETTE_SIZE];
+extern const uint8_t FONT_SANS_PALETTE[FONT_SANS_PALETTE_SIZE];
